backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Configuration and validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0

# Authentication & Encryption
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# CORS
python-cors==1.0.0

# RAG and embeddings
langchain==0.1.0
langchain-community==0.0.10
langchain-chroma==0.1.0
chromadb==0.4.22
sentence-transformers==2.3.1
docling==1.0.0

# GitHub integration
PyGithub==2.1.1
